# springboot_openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
paths:
  /suspensions/byUser/{username}/last/isRejected:
    put:
      tags:
      - suspension-controller
      operationId: setUserLastSuspensionRejectedStatus
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DAOSuspensionRejectStatus"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /race:
    post:
      tags:
      - race-controller
      operationId: newRace
      parameters:
      - name: vehicleId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: username
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: date-time
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "405":
          description: Vehicle doesn't exist
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /race/end:
    post:
      tags:
      - race-controller
      operationId: endRace
      parameters:
      - name: vehicleId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: username
        in: query
        required: true
        schema:
          type: string
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
        "402":
          description: User is suspended
          content:
            application/text:
              schema: {}
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "409":
          description: Vehicle not parked in the dock
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /suspensions:
    get:
      tags:
      - suspension-controller
      operationId: getAll_1
      parameters:
      - name: active
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOCustomerWithSuspension"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /suspensions/byUser/{username}/suspended-permanently:
    get:
      tags:
      - suspension-controller
      operationId: isPermanentSuspended
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /customers/{username}/races:
    get:
      tags:
      - customer-controller
      operationId: getRacesByUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAORaceOnly"
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /customers/{username}/credit:
    get:
      tags:
      - customer-controller
      operationId: getCreditByCustomer
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
                format: float
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
components:
  schemas:
    DAOSuspensionRejectStatus:
      required:
      - isRejected
      type: object
      properties:
        isRejected:
          type: boolean
...
    DAOCustomerWithSuspension:
      required:
      - email
      - name
      - paid
      - start
      - surname
      - username
      - value
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        name:
          type: string
        surname:
          type: string
        start:
          type: integer
          format: int64
        end:
          type: integer
          format: int64
        value:
          type: number
          format: float
        paid:
          type: boolean
...
    DAORaceOnly:
      required:
      - customer
      - id
      - minDuration
      type: object
      properties:
        id:
          $ref: "#/components/schemas/RaceId"
        customer:
          type: string
        minDuration:
          type: integer
          format: int32
    RaceId:
      required:
      - idCustomer
      - idVehicle
      - start
      type: object
      properties:
        start:
          type: string
          format: date-time
        idCustomer:
          maxLength: 50
          minLength: 0
          type: string
        idVehicle:
          type: integer
          format: int32